requirements.txt
# Core ML
pandas==2.3.3
numpy==2.1.3
scikit-learn==1.7.2
# AutoGluon (explicitly specify all models used in training)
autogluon.tabular==1.4.0
lightgbm
catboost
xgboost
torch
fastai
pyarrow==20.0.0

# MLOps
dvc==3.64.2
mlflow==3.7.0
psycopg2-binary==2.9.11
evidently==0.7.17
wandb==0.23.1
requests==2.32.5

# API (Phase 5)
fastapi==0.124.4
uvicorn[standard]==0.38.0
pydantic==2.12.5
apscheduler==3.10.4

# Config
python-dotenv==1.2.1
